webapi/webapi-webrtc-w3c-tests/webrtc/blink/RTCPeerConnection-localDescription.html

<!DOCTYPE html>
<!--
Test converted from blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/mediastream/RTCPeerConnection-localDescription.html
-->

<meta charset=utf-8>
<title>WebRTC Test: RTCPeerConnection localDescription</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel=help href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#widl-RTCPeerConnection-setLocalDescription-void-RTCSessionDescription-description-VoidFunction-successCallback-RTCPeerConnectionErrorCallback-failureCallback">
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<div id=log></div>
<script>

var t = async_test(document.title);

var pc = null;
var pc2 = null;
var atStep = "Create offer at one";

function requestSucceeded1 (offer) {
  t.step(function() {
    atStep = 'Set local description at one';
    pc.setLocalDescription(offer)
    .then(function() {
      if (typeof RTCPeerConnection != "undefined") {
        pc2 = new RTCPeerConnection(null, null);
      } else {
        pc2 = new webkitRTCPeerConnection(null, null);
      }
      atStep = 'Set remote description at two';
      return pc2.setRemoteDescription(offer);
    }).then(function() {
      atStep = 'Create answer at two';
      pc2.createAnswer(requestSucceeded2, requestFailed2);
    }).catch(t.step_func(function(e) {
      assert_unreached('Error ' + e.name + ': ' + e.message +
                       ' happened at step ' + atStep);
    }));
  });
}

function requestFailed1 () {
  t.step(function () {
    assert_unreached("requestFailed1 was called");
  });
  t.done();
}

function requestSucceeded2 (answer) {
  t.step(function() {
    atStep = 'Set local description at two';
    pc2.setLocalDescription(answer)
    .then(function() {
      t.done();
    }).catch(t.step_func(function(e) {
      assert_equals(pc.localDescription.type, "offer");
      assert_equals(pc.localDescription.sdp, "local");
      pc.close();
      assert_equals(pc.localDescription.type, "offer");
      assert_equals(pc.localDescription.sdp, "local");
      assert_unreached('Error ' + e.name + ': ' + e.message +
                       ' happened at step ' + atStep);
    }));
  });
}

function requestFailed2 () {
  t.step(function () {
    assert_unreached("requestFailed4 was called");
  });
  t.done();
}

t.step(function() {
  if (typeof RTCPeerConnection != "undefined") {
    pc = new RTCPeerConnection(null, null);
  } else {
    pc = new webkitRTCPeerConnection(null, null);
  }
  pc.createOffer(requestSucceeded1, requestFailed1);
});

</script>
